Refactor imports and add EmbeddingsHandler class
- Cleaned up import statements across multiple files for consistency and clarity.
- Introduced a new `EmbeddingsHandler` class in `get_embeddings.py` to manage embedding functionalities.
- Updated references in `yt_fts.py`, `search.py`, and `chatbot.py` to align with the new import structure.
